requirements.txt
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.0
idna==3.4
imageio==2.26.0
kiwisolver==1.4.4
lazy_loader==0.1
matplotlib==3.7.1
networkx==3.0
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.28.2
scikit-image==0.20.0
scipy==1.10.1
six==1.16.0
tifffile==2023.2.28
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
